For the stop bits getter and setter methods, use vpr::Uint8 instead of vpr::Uint16.  16 bits is far more than necessary.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@4819 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/vapor/vpr/IO/Port/SerialPort_t.h
@@ -307,14 +307,14 @@ public:
     // ------------------------------------------------------------------------
     // ------------------------------------------------------------------------
     inline vpr::Status
-    getStopBits (vpr::Uint16& num_bits) {
+    getStopBits (vpr::Uint8& num_bits) {
         return m_sio_imp.getStopBits(num_bits);
     }
 
     // ------------------------------------------------------------------------
     // ------------------------------------------------------------------------
     inline vpr::Status
-    setStopBits (const vpr::Uint16 num_bits) {
+    setStopBits (const vpr::Uint8 num_bits) {
         return m_sio_imp.setStopBits(num_bits);
     }
